+++ src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="app">
     <router-view/>
-    <FootGuide/>
+    <FootGuide v-show="$route.meta.showFoot"/>
   </div>
 </template>
