+++ src/components/ShopList/images/stars/star24_on@2x.png
