+++ src/components/ShopList/images/stars/star48_half@3x.png
