+++ src/router/index.js
@@ -1,35 +1,53 @@
-/*路由器对象模块*/
-
+/*
+路由器对象模块
+ */
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 
-import MSite from '../pages/MSite/MSite'
-import Search from '../pages/Search/Search'
-import Order from '../pages/Order/Order'
-import Profile from '../pages/Profile/Profile'
+import MSite from '../pages/MSite/MSite.vue'
+import Search from '../pages/Search/Search.vue'
+import Order from '../pages/Order/Order.vue'
+import Profile from '../pages/Profile/Profile.vue'
+import Login from '../pages/Login/Login.vue'
 
 // 声明使用vue插件
 Vue.use(VueRouter)
 
 export default new VueRouter({
   mode: 'history',
-  // 应用中所有的路由
+  // 应用中所有路由
   routes: [
     {
       path: '/msite',
-      component: MSite
+      component: MSite,
+      meta: {
+        showFoot: true
+      }
     },
     {
       path: '/search',
-      component: Search
+      component: Search,
+      meta: {
+        showFoot: true
+      }
     },
     {
       path: '/order',
-      component: Order
+      component: Order,
+      meta: {
+        showFoot: true
+      }
     },
     {
       path: '/profile',
-      component: Profile
+      component: Profile,
+      meta: {
+        showFoot: true
+      }
+    },
+    {
+      path: '/login',
+      component: Login
     },
     {
       path: '/',
